+++ DbUtils/Utils/MySqlUtils.cs
@@ -27,7 +27,7 @@ namespace DbUtils.Utils
             public string IS_NULLABLE { get; set; }
             public string COLUMN_TYPE { get; set; }
             public string COLUMN_COMMENT { get; set; }
-            public int? CHARACTER_MAXIMUM_LENGTH { get; set; }
+            public ulong? CHARACTER_MAXIMUM_LENGTH { get; set; }
             public int? NUMERIC_PRECISION { get; set; }
             public int? NUMERIC_SCALE { get; set; }
         }
